% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymeinhibition.r
\name{SScompinh}
\alias{SScompinh}
\title{Supplies a selfStart function for modeling competitive inhibition of
  enzyme kinetics.}
\usage{
SScompinh(sub, inh, Vmax, Km, Kc)
}
\arguments{
\item{sub}{A vector of substrate concentrations.}

\item{inh}{A vector of inhibitor concentrations.}

\item{Vmax}{The maximum initial reaction rate.}

\item{Km}{The Michaelis constant for the enzyme-substrate complex.}

\item{Kc}{The dissociation constant for the enzyme-inhibitor complex.}
}
\value{
A \code{\link{selfStart}} function for use in nonlinear
          optimization.
}
\description{
Supplies a selfStart function for modeling competitive inhibition of
  enzyme kinetics.
}
\examples{
\dontrun{compnls <- nls(Rate~SScompinh(Sub.Con,Inh.Con,Vm,Km,Kcomp),
                        data=Enzymeframe)}
}
\seealso{
\code{\link{selfStart}}, \code{\link[stats]{nls}}
}
