% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymeinhibition.r
\docType{package}
\name{SSenzymekinetics-package}
\alias{SSenzymekinetics}
\alias{SSenzymekinetics-package}
\title{Model different types of enzyme kinetic behaviors for nonlinear
  optimization of rate data.}
\description{
The selfStart functions provided by the package are
\itemize{
  \item \code{\link{SScompinh}} Model for competitive inhibition.
  \item \code{\link{SSuncompinh}} Model for uncompetitive inhibition.
  \item \code{\link{SSmixedinh}} Model for mixed inhibition.
}
}
\details{
The functions are intended to be used as the right-hand side of
formulas in \code{\link[stats]{nls}} or similar functions. See the
documentation of the functions for examples.
}
\author{
\strong{Maintainer}: Nathan Malmberg \email{njmalmberg@gmail.com}

}
