% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymeinhibition.r
\name{SSmixedinh}
\alias{SSmixedinh}
\title{Supplies a selfStart function to model mixed inhibition of enzyme
  kinetics.}
\usage{
SSmixedinh(sub, inh, Vmax, Km, Kc, Ku)
}
\arguments{
\item{sub}{A vector of substrate concentrations.}

\item{inh}{A vector of inhibitor concentrations.}

\item{Vmax}{The maximum initial rate of reaction.}

\item{Km}{The Michaelis constant for the enzyme-substrate complex.}

\item{Kc}{The dissociation constant for the enzyme-inhibitor complex.}

\item{Ku}{The dissociation constant for the ternary complex.}
}
\value{
A \code{\link{selfStart}} function for use in nonlinear
         optimization.
}
\description{
Supplies a selfStart function to model mixed inhibition of enzyme
  kinetics.
}
\examples{
\dontrun{mixed <- nls(Rate~SSmixedinh(Sub.Con,Inh.Con,Vm,Km,Kc,Ku),
                      data=EnzymeFrame)}
}
\seealso{
\code{\link[stats]{selfStart}}, \code{\link[stats]{nls}}
}
