% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymeinhibition.r
\name{SSuncompinh}
\alias{SSuncompinh}
\title{Supplies a selfStart function for modeling uncompetitive inhibition of
  enzyme kinetics.}
\usage{
SSuncompinh(sub, inh, Vmax, Km, Ku)
}
\arguments{
\item{sub}{A vector of substrate concentrations.}

\item{inh}{A vector of inhibitor concentrations.}

\item{Vmax}{The maximum initial reaction rate.}

\item{Km}{The Michaelis constant for the enzyme-substrate complex.}

\item{Ku}{The dissociation constant for the ternary complex.}
}
\value{
A \code{\link{selfStart}} function for use in nonlinear
          optimization.
}
\description{
Supplies a selfStart function for modeling uncompetitive inhibition of
  enzyme kinetics.
}
\examples{
\dontrun{x <- nls(Rate~SSuncompinh(Sub.Con,Inh.Con,Vm,Km,Ku),
                  data=EnzymeFrame)}
}
\seealso{
\code{\link[stats]{selfStart}}, \code{\link[stats]{nls}}.
}
